requirements.txt
Django==5.0.1
python-dotenv==1.0.0
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.1.0
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-compressor==4.4
django-redis==5.4.0
psycopg2-binary==2.9.9
cent==4.1.0
django-storages==1.14.2
django-allauth==0.59.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
